Refactor: 폴더 구조 수정 - 최소화 된 feature 기반
