+++ src/styles/step-styles.ts
